refactor(book-list): migrate book detail and content list
This commit also remove old and unused components from previous routes
