--- app/(protected)/(bookList)/dialog/content-form/constants.ts
@@ -1,4 +1,0 @@
-export const label = {
-  bookId: "Buku",
-  path: "Link",
-} as const;
