--- app/(protected)/(bookList)/dialog/content-form/index.ts
@@ -1,3 +1,0 @@
-export * from "./content-form";
-export * from "./handler";
-export * from "./constants";
