+++ app/(protected)/(books)/[bookId]/@list/components/toolbar.tsx
@@ -80,7 +80,7 @@ export const Toolbar = () => {
         <Button
           variant="outline"
           size="icon"
-          className="rounded-full flex-shrink-0"
+          className="rounded-full flex-shrink-0 gap-1 size-9 lg:size-fit border text-sky-500 hover:text-sky-600 border-sky-500 hover:border-sky-600 dark:border-sky-800 hover:bg-sky-300/20 dark:hover:bg-sky-700/20"
           onClick={onFilterClick}
         >
           <i className="bx bx-filter-alt text-sm md:text-lg text-primary-500-400-token" />
@@ -91,7 +91,7 @@ export const Toolbar = () => {
       </div>
       <Button
         size="icon"
-        className="rounded-full flex-shrink-0"
+        className="rounded-full flex-shrink-0 gap-1 size-9 lg:size-fit bg-sky-500 dark:bg-sky-900 hover:bg-sky-600 dark:hover:bg-sky-700 text-white dark:text-white hover:text-white"
         onClick={() => {
           dialog &&
             dialog.openDialog("form", EMPTY_BOOK_TEMPLATE, async (result) => {
