+++ app/(protected)/(books)/[bookId]/components/attribute-list.tsx
