+++ app/(protected)/(books)/[bookId]/components/handler.ts
