+++ app/(protected)/(books)/[bookId]/components/no-match-search.tsx
@@ -1,8 +1,12 @@
-"use client";
-
 import Image from "next/image";
 
-export const NoMatchSearch = () => {
+export const NoMatchSearch = ({
+  searchParams,
+}: {
+  searchParams: Record<string, string | string[] | undefined>;
+}) => {
+  const serachQuery = searchParams["content"] as string;
+
   return (
     <div className="flex-grow flex items-center justify-center">
       <div className="rounded-lg p-4 bg-surface-50-900-token flex-1 flex flex-col items-center justify-center my-auto gap-6">
@@ -15,6 +19,10 @@ export const NoMatchSearch = () => {
         </div>
         <div className="space-2 text-center max-w-xs">
           <p className="font-medium">Konten digital yang ada cari tidak ada</p>
+          <p className="text-xs text-zinc-700 dark:text-zinc-200 opacity-60 mt-1">
+            Belum terdapat konten digital yang sesuai dengan kata kunci:{" "}
+            {`"${serachQuery}"`}
+          </p>
         </div>
       </div>
     </div>
